// app/interfaces/interface.tsx
export interface RoomCount{
  idRoomCount: number;
  roomCountName: string;
};

export interface BuildingType  {
  idBuildingType: number;
  buildingTypeName: string;
};
export interface PropertyType {
  idPropertyType: number;
  propertyTypeName: string;
};

export interface MetroType{
  id: number;
  metroName: string;
};


export interface RegionType {
  idRegion: number;
  regionName: string;
};